% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/birth.R
\name{birth}
\alias{birth}
\title{LINKAGES birth function}
\usage{
birth(nspec, ntrees, frt, iage, slta, sltb, dbh, fwt, switch.mat, degd, dmin,
  dmax, frost, rt, itol, mplant, nogro, ksprt, sprtnd, max.ind, smgf, degdgf)
}
\arguments{
\item{nspec}{number of species}

\item{ntrees}{number of trees of each species}

\item{frt}{foliage retention time}

\item{iage}{age of each individual}

\item{slta}{parameter to calculate crown area from diameter}

\item{sltb}{parameter to calculate crown area from diameter}

\item{dbh}{diameter of each individual}

\item{fwt}{leaf weight per unit crown area}

\item{switch.mat}{TRUE/FALSE switches for each species}

\item{degd}{growing degree days from tempe.R}

\item{dmin}{degree day minimum for each species}

\item{dmax}{degree day maximum for each species}

\item{frost}{minimum January temperature tolerated}

\item{rt}{current monthly temperatures}

\item{itol}{shade tolerance code for each species}

\item{mplant}{maximum seeding in rate per plot}

\item{nogro}{flags individuals that aren't growing well}

\item{ksprt}{flags individuals that could sprout}

\item{sprtnd}{number of sprouts per stump}

\item{max.ind}{maximum number of individuals in a plot}

\item{smgf}{soil moisture growing factor from gmult.R}

\item{degdgf}{growing degree day growing factor from gmult.R}
}
\value{
iage age of each individual

dbh diameter of each individual

nogro flags individuals growing slowly

ntrees number of trees of each species

newtr records species considered 'eligible'

ksprt flags individuals that could sprout
}
\description{
BIRTH CALCULATES SEEDLING AND SPROUT BIRTH BASED ON
   SPECIES FECUNDITY, SEEDBED CONDITIONS, SUSCEPTIBILITY TO BROWSING,
   AND THE DEGREE TO WHICH LIGHT, SOIL MOISTURE, AND DEGREE DAYS ARE
   LESS THAN OPTIMUM FOR GROWTH. SOIL MOISTURE AND DEGREE DAY
   MULTIPLIERS ARE SUPPLIED BY SUBROUTINE GMULT.
   A SPECIES CAN HAVE SPROUTS IF AT LEAST ONE
   TREE WITH DIAMETER BETWEEN SPRTMN AND SPRTMX DIED LAST
   YEAR (KSPRT INCREMENTED BY 1 IN KILL).
   RANDOM NUMBERS USED TO DETERMINE OCCURENCE OF BROWSING, NUMBERS
   OF SEEDLINGS AND SPROUTS, AND DBH SUPPLIED BY rand.
}
\author{
Ann Raiho
}

