% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decomp.R
\name{decomp}
\alias{decomp}
\title{LINKAGES decomp function}
\usage{
decomp(fdat, aet, ncohrt, fc, dry, tyl, C.mat)
}
\arguments{
\item{fdat}{current year}

\item{aet}{average temperature by month for current year}

\item{ncohrt}{average precipitation by month for current year}

\item{fc}{field capacity}

\item{dry}{wilting point}

\item{tyl}{beginning growing season day of year}

\item{C.mat}{end growing season day of year}
}
\value{
ff weights and n content of forest floor by litter type

availn available N

tyln leaf litter N content

hcn humus C:N ratio

sco2 total soil co2-c

ncohrt number of cohorts

C.mat matrix for data on litter cohorts
}
\description{
DECOMP CALCULATES CARBON AND NITROGEN FLOWS THROUGH
   SOIL. AVAILABLE N (AVAILN) IS USED IN GMULT TO CALCULATE
  SOIL NITROGEN GROWTH MULTIPLIERS. AET IS FED IN FROM MOIST.
   THIS YEAR'S LEAF, TWIG, ROOT, AND WOOD LITTER IS FED IN FROM KILL
  (ARRAY TYL). THE SIMULATION STARTS ON BARE GROUND (ONLY HUMUS
  PRESENT. BASESC AND BASESN ARE STARTING HUMUS WEIGHT AND N
  CONTENTS READ IN INPUT). THREE TYPES OF SOIL ORGANIC MATTER ARE
   RECOGNIZED: COHORTS EITHER IMMOBILIZING OR RAPIDLY MINERALIZING
   NITROGEN AND A HOMOGENOUS HUMUS POOL SLOWLY MINERALIZING N.
}
\author{
Ann Raiho
}

