% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmult.R
\name{gmult}
\alias{gmult}
\title{LINKAGES gmult function}
\usage{
gmult(egs, bgs, availn, degd, dmin, dmax, d3, fj, cm1, cm3, cm2, cm4, cm5,
  nspec)
}
\arguments{
\item{egs}{end of growing season day of year}

\item{bgs}{beginning of growing season day of year}

\item{availn}{available nitrogen}

\item{degd}{total growing degree days}

\item{dmin}{degree day minimum for each species}

\item{dmax}{degree day maximum for each species}

\item{d3}{drought tolerance (fraction of the growing season)}

\item{fj}{number of dry days}

\item{cm1}{parameter for nitrogen growth multipliers}

\item{cm3}{parameter for nitrogen growth multipliers}

\item{cm2}{parameter for nitrogen growth multipliers}

\item{cm4}{parameter for nitrogen growth multipliers}

\item{cm5}{parameter for nitrogen growth multipliers}

\item{nspec}{number of species}
}
\value{
smgf soil moisture growth factor

sngf soil nitrogen growth factor

degdgf growing degree day growth factor

availn available nitrogen
}
\description{
GMULT CALCULATES DEGREE DAY, SOIL MOISTURE, AND SOIL
                NITROGEN MULTIPLIERS USED IN SUBROUTINES BIRTH AND GROW BASED
                ON ON DEGD (SUPPLIED BY TEMPE), FJ (SUPPLIED BY MOIST), AND
                AVAILN (SUPPLIED BY DECOMP), RESPECTIVELY.
}
\author{
Ann Raiho
}

