% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow.R
\name{grow}
\alias{grow}
\title{LINKAGES growth function}
\usage{
grow(max.ind, nspec, ntrees, frt, slta, sltb, dbh, fwt, b2, b3, itol, g, degdgf,
  smgf, sngf, frost, rt, iage, nogro)
}
\arguments{
\item{max.ind}{maximum number of individuals}

\item{nspec}{number of species}

\item{ntrees}{number of trees of each species}

\item{frt}{foliage retention time}

\item{slta}{parameter to calculate crown area from diameter}

\item{sltb}{parameter to calculate crown area from diameter}

\item{dbh}{diameter of each tree}

\item{fwt}{leaf weight per unit crown area}

\item{b2}{growth scaling parameter}

\item{b3}{growth scaling parameter}

\item{itol}{shade tolerance code for each species}

\item{g}{scalar for species maximum diameter increment}

\item{degdgf}{growing degree day growth factor from gmult.R}

\item{smgf}{soil moisture growth factor from gmult.R}

\item{sngf}{soil nitrogen growth factor from gmult.R}

\item{frost}{minimum January temperature tolerated}

\item{rt}{current monthly temperature}

\item{iage}{age of each individual}

\item{nogro}{flags slow growing individuals}
}
\value{
dbh diameter of each individual

ntrees number of trees of each species

awp aboveground woody production
}
\description{
GROW CALCULATES DIAMETER GROWTH FOR EACH TREE
  BY DECREASING MAXIMAL GROWTH TO THE EXTENT THAT THE MOST
  LIMITING RESOURCE IS LESS THAN OPTIMAL.
}
\author{
Ann Raiho
}

