% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkages.R
\name{linkages}
\alias{linkages}
\title{LINKAGES main function}
\usage{
linkages(iplot, nyear, nspec, fc, dry, bgs, egs, max.ind, plat)
}
\arguments{
\item{iplot}{PEcAn settings object}

\item{nyear}{data.frame of initial condition sample (nens X nstate)}

\item{nspec}{data.frame of model parameter sample (nense X nstate)}

\item{fc}{data.frame of observations with columns: mean, sd}

\item{dry}{data.frame of observations with columns: mean, sd}

\item{bgs}{data.frame of observations with columns: mean, sd}

\item{egs}{data.frame of observations with columns: mean, sd}

\item{max.ind}{data.frame of observations with columns: mean, sd}

\item{plat}{data.frame of observations with columns: mean, sd}
}
\value{
several things ---> should this be different
}
\description{
Main function for running all LINKAGES subroutines
}
\author{
Ann Raiho
}

