% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moist.R
\name{moist}
\alias{moist}
\title{LINKAGES moist function}
\usage{
moist(kyr, temp.vec, precip.vec, fc, dry, bgs, egs, plat, clat)
}
\arguments{
\item{kyr}{current year}

\item{temp.vec}{average temperature by month for current year}

\item{precip.vec}{average precipitation by month for current year}

\item{fc}{field capacity}

\item{dry}{wilting point}

\item{bgs}{beginning growing season day of year}

\item{egs}{end growing season day of year}

\item{plat}{plot latititude}

\item{clat}{climate adjustment matrix}
}
\value{
aet             actual evapotranspiration (units:mm)

fj              total number of dry days
}
\description{
Calculates actual evapotranspiration (AET) for the current year
}
\author{
Ann Raiho
}

