% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotin.R
\name{plotin}
\alias{plotin}
\title{LINKAGES plotin function}
\usage{
plotin(iplot, basesc, basesn, max.ind, nspec)
}
\arguments{
\item{iplot}{number of plots}

\item{basesc}{initial humus weight}

\item{basesn}{initial humus nitrogen content}

\item{max.ind}{maximum number of individuals}

\item{nspec}{number of species in simulation}
}
\value{
ntrees             matrix for number of trees for each species

dbh                matrix for diameter at breast height for each tree

nogro              used to flag slow growing trees

ksprt              used to flag trees eligible to sprout

iage               matrix for age of each tree

C.mat              matrix for data on litter cohorts

ncohrt             number of cohorts

tyl                total yearly litter
}
\description{
Sets up initial conditions for LINAKGES
}
\author{
Ann Raiho
}

